+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Private function:
+  - `Invoke-ReportServerSetupAction`
+
 ### Changed
 
 - SqlServerDsc